Config Return Data

+++ controllers/handleFactory.js
@@ -1,15 +1,23 @@
 import catchAsync from "../utils/catchAsync.js";
 import AppError from "../utils/appError.js";
 import APIFeatures from "../utils/apiFeatures.js";
+import customResourceResponse from "../utils/constant.js";
 
 export const createOne = (Model) =>
   catchAsync(async (req, res, next) => {
     const document = await Model.create(req.body);
-    res.status(200).json({
+    if (!document) {
+      return next(
+        new AppError(
+          customResourceResponse.serverError.message,
+          customResourceResponse.serverError.statusCode
+        )
+      );
+    }
+    res.status(customResourceResponse.created.statusCode).json({
       status: "success",
-      data: {
-        data: document,
-      },
+      message: customResourceResponse.created.message,
+      data: { data: doc._id },
     });
   });
 
@@ -18,20 +26,29 @@ export const deleteOne = (Model) =>
     const id = req.params.id;
 
     // ///Check ID is valid
-    // if (!id.match(/^[0-9a-fA-F]{24}$/)) {
-    //   return next(new AppError('Id is not valid', 404));
-    // }
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+      return next(
+        new AppError(
+          customResourceResponse.notValidId.message,
+          customResourceResponse.notValidId.statusCode
+        )
+      );
+    }
 
     const document = await Model.findByIdAndDelete(id);
 
     if (!document) {
-      return next(new AppError(`No document found with that ID`, 404));
+      return next(
+        new AppError(
+          customResourceResponse.recordNotFoundOne.message,
+          customResourceResponse.recordNotFoundOne.statusCode
+        )
+      );
     }
-    res.status(200).json({
+    res.status(customResourceResponse.success.statusCode).json({
       status: "success",
-      data: {
-        data: null,
-      },
+      message: customResourceResponse.success.message,
+      data: { data: doc },
     });
   });
 
@@ -40,9 +57,14 @@ export const updateOne = (Model) =>
     const id = req.params.id;
 
     // ///Check ID is valid
-    // if (!id.match(/^[0-9a-fA-F]{24}$/)) {
-    //   return next(new AppError('Id is not valid', 404));
-    // }
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+      return next(
+        new AppError(
+          customResourceResponse.notValidId.message,
+          customResourceResponse.notValidId.statusCode
+        )
+      );
+    }
 
     const document = await Model.findByIdAndUpdate(
       id,
@@ -54,13 +76,17 @@ export const updateOne = (Model) =>
     );
 
     if (!document) {
-      return next(new AppError(`No document found with that ID`, 404));
+      return next(
+        new AppError(
+          customResourceResponse.recordNotFoundOne.message,
+          customResourceResponse.recordNotFoundOne.statusCode
+        )
+      );
     }
-    res.status(200).json({
+    res.status(customResourceResponse.success.statusCode).json({
       status: "success",
-      data: {
-        data: document,
-      },
+      message: customResourceResponse.success.message,
+      data: { data: doc },
     });
   });
 
@@ -69,9 +95,14 @@ export const getOne = (Model, popOptions) =>
     const id = req.params.id;
 
     // ///Check ID is valid
-    // if (!id.match(/^[0-9a-fA-F]{24}$/)) {
-    //   return next(new AppError('Id is not valid', 404));
-    // }
+    if (!id.match(/^[0-9a-fA-F]{24}$/)) {
+      return next(
+        new AppError(
+          customResourceResponse.notValidId.message,
+          customResourceResponse.notValidId.statusCode
+        )
+      );
+    }
 
     ///Return join collection child references and auto parse value according to _id
     const document = Model.findById(id);
@@ -84,11 +115,17 @@ export const getOne = (Model, popOptions) =>
     // await tour.findById(id).populate('guides');
     // console.log(tourDetail);
     // await tour.findOne({ _id: id });
-    if (!tourDetail) {
-      return next(new AppError("No document found with that ID", 404));
+    if (!doc) {
+      return next(
+        new AppError(
+          customResourceResponse.recordNotFoundOne.message,
+          customResourceResponse.recordNotFoundOne.statusCode
+        )
+      );
     }
-    res.status(200).json({
+    res.status(customResourceResponse.success.statusCode).json({
       status: "success",
+      message: customResourceResponse.success.message,
       data: { data: doc },
     });
   });
@@ -108,7 +145,16 @@ export const getAll = (Model) =>
     const doc = await features.query;
 
     // SEND RESPONSE
-    res.status(200).json({
+    if (!doc) {
+      return next(
+        new AppError(
+          customResourceResponse.recordNotFound.message,
+          customResourceResponse.recordNotFound.statusCode
+        )
+      );
+    }
+    res.status(customResourceResponse.success.statusCode).json({
+      message: customResourceResponse.success.message,
       status: "success",
       results: doc.length,
       data: {
